backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<3.15
psycopg2-binary>=2.9,<3.0 # PostgreSQL adapter
django-environ>=0.11,<0.12 # For reading .env file easily
celery>=5.3,<5.4
redis>=5.0,<5.1 # Python client for Redis (needed by Celery)
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=3.14,<4.0
Pillow>=10.0,<11.0 # For image processing
mutagen>=1.47
django-filter>=24.2
pydub>=0.25